Reorganize gerrit listener
The listener process was in its own directory to isolate it from everything
else we have in this repo. It doesn't necessarily need to be in its own
location, and it would make sense to have the scheduler process live in the
same place. This change renaming the location of where the listener process
lives in preparing for the scheduler.

It also adds a bit of information in a separate README.md about how the two
processes work together to process performance patches.