Remove concurrency performance tests
When the benchmark.sh script was originally written, it was used to also test
the tuning of the web server, hence the need for concurrency tests. Since the
OSIC Performance Bot is mainly focused on capturing only keystone performance,
we should be able to get that data from non-concurrent test runs.

This commit removes the concurrent tests from the benchmark script. As a
result, this will quicken the test runs a bit.

Fixes #4