Add script to populate keystone with fake data
The `bootstrap.py` script takes a stab at populating keystone with some data
before running any sort of performance test. This is an attempt to model some
sort of real deployment and make it so that our performance numbers aren't
terribly unrealistic because keystone doesn't have enough data. An example of
this would be making sure we have enough endpoints and services in the catalog
instead of just a single service and a single endpoint.

Fixes #6